Add notification above seekbar
