+++ src/components/song.js
@@ -48,7 +48,7 @@ export const SongItem = ({
         <IconPress
           onPress={() =>
             shareUrl(
-              `Şubadap'tan ${song.name} şarkısını dinle: https://subadapp.page.link/?no=${song.no}. Şubadapp uygulamasını indir: ${PLAY_STORE_URL}`
+              `Şubadap'tan ${song.name} şarkısını dinle: https://subadapp.page.link/?song=${song.no}. Şubadapp uygulamasını indir: ${PLAY_STORE_URL}`
             )
           }
           icon={faShare}
